trains/requirements.frozen.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.frozen.txt --strip-extras requirements.txt
#
aiofiles==24.1.0
    # via nicegui
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via
    #   -r requirements.txt
    #   nicegui
    #   python-socketio
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   nicegui
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
docutils==0.21.2
    # via nicegui
fastapi==0.115.12
    # via nicegui
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
geojson-pydantic==1.2.0
    # via -r requirements.txt
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via nicegui
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ifaddr==0.2.0
    # via nicegui
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via nicegui
jinja2==3.1.6
    # via nicegui
markdown2==2.5.3
    # via nicegui
markupsafe==3.0.2
    # via jinja2
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
mypy==1.15.0
    # via -r requirements.txt
mypy-extensions==1.1.0
    # via mypy
nicegui==2.15.0
    # via -r requirements.txt
numpy==2.2.5
    # via shapely
orjson==3.10.16
    # via nicegui
packaging==25.0
    # via pytest
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pscript==0.7.7
    # via vbuild
pydantic==2.11.3
    # via
    #   -r requirements.txt
    #   fastapi
    #   geojson-pydantic
pydantic-core==2.33.1
    # via pydantic
pygments==2.19.1
    # via nicegui
pytest==8.3.5
    # via
    #   -r requirements.txt
    #   pytest-asyncio
pytest-asyncio==0.26.0
    # via -r requirements.txt
python-dotenv==1.1.0
    # via uvicorn
python-engineio==4.12.0
    # via
    #   nicegui
    #   python-socketio
python-multipart==0.0.20
    # via nicegui
python-socketio==5.13.0
    # via nicegui
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   -r requirements.txt
    #   nicegui
shapely==2.1.0
    # via -r requirements.txt
simple-websocket==1.1.0
    # via python-engineio
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via
    #   fastapi
    #   nicegui
types-requests==2.31.0.6
    # via -r requirements.txt
types-urllib3==1.26.25.14
    # via types-requests
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   mypy
    #   nicegui
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==1.26.20
    # via
    #   -r requirements.txt
    #   nicegui
    #   requests
uvicorn==0.34.2
    # via nicegui
uvloop==0.21.0
    # via uvicorn
vbuild==0.8.2
    # via nicegui
watchfiles==1.0.5
    # via
    #   nicegui
    #   uvicorn
websockets==15.0.1
    # via uvicorn
wsproto==1.2.0
    # via simple-websocket
yarl==1.20.0
    # via aiohttp
